Añadido ColorPicking y mejorado el UUID con funcion de busqueda de entidades por UUID

+++ Core/src/core/UUID.h
@@ -7,34 +7,33 @@
 
 namespace libCore {
 
-	class UUID
-	{
-	public:
-		UUID();
-		UUID(uint64_t uuid);
-		UUID(const std::string& uuidStr); // Constructor desde cadena
-		UUID(const UUID&) = default;
+    class UUID
+    {
+    public:
+        UUID();
+        UUID(uint32_t uuid);  // Cambiar a 24 bits
+        UUID(const std::string& uuidStr); // Constructor desde cadena
+        UUID(const UUID&) = default;
 
-		operator uint64_t() const { return m_UUID; }
+        operator uint32_t() const { return m_UUID; }  // Cambiar a 24 bits
 
-		std::string ToString() const; // Convertir UUID a cadena
+        std::string ToString() const; // Convertir UUID a cadena
 
-	private:
-		uint64_t m_UUID;
-	};
+    private:
+        uint32_t m_UUID;  // Cambiar a 24 bits
+    };
 
 }
 
 namespace std {
-	template <typename T> struct hash;
+    template <typename T> struct hash;
 
-	template<>
-	struct hash<libCore::UUID>
-	{
-		std::size_t operator()(const libCore::UUID& uuid) const
-		{
-			return (uint64_t)uuid;
-		}
-	};
-
+    template<>
+    struct hash<libCore::UUID>
+    {
+        std::size_t operator()(const libCore::UUID& uuid) const
+        {
+            return (uint32_t)uuid;  // Cambiar a 24 bits
+        }
+    };
 }
